Updated with advanced CVE scanning
Permissions added to yaml to ensure security by default. A separate tool is run to checkout the repo and read the code. Load true is set to export the image to docker images, allowing the trivy scans to make use of docker commands. The advanced scan then uploads the results of any vulnerabilities in the code to github security